refactor: resolve SonarCloud duplication failures in ViewModels
